+++ src/Lib/Session.php
@@ -514,7 +514,11 @@ class Session
 			$_SESSION['client']['betriebe'] = array();
 			foreach ($r as $rr) {
 				// add info about the next free pickup slot to the store
-				$rr['pickupStatus'] = $this->storeService->getAvailablePickupStatus($rr['id']);
+
+				//temporarily disable pickup slot markers for production release
+				//$rr['pickupStatus'] = $this->storeService->getAvailablePickupStatus($rr['id']);
+				$rr['pickupStatus'] = 0; // STATUS_GREEN (no color) = 0
+
 				$_SESSION['client']['betriebe'][$rr['id']] = $rr;
 			}
 		}